fix: removed types that broke the build
