+++ src/components/ui/header.tsx
@@ -1,5 +1,7 @@
-
 "use client";
+
+import { useCallback } from "react";
+
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import {
   DropdownMenu,
@@ -12,15 +14,16 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { useNostrContext } from "@/lib/nostr/NostrContext";
 import useSession from "@/lib/nostr/useSession";
-import { ChevronDown } from "lucide-react";
-import { HeaderConfig } from "types";
-import { MainNav } from "../main-nav";
-import { Button } from "./button";
-import { useRouter } from "next/navigation";
-import { useCallback } from "react";
-import Link from "next/link";
 
-const HeaderConfig: HeaderConfig = {
+import { ChevronDown } from "lucide-react";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
+
+import { MainNav } from "../main-nav";
+
+import { Button } from "./button";
+
+const HeaderConfig = {
   mainNav: [
     {
       title: "Pull Requests",
@@ -92,14 +95,13 @@ const PrimaryGitInfo = DropdownItems.slice(0, 8);
 const restGitInfo = DropdownItems.slice(8);
 
 export function Header() {
-
-  const {picture, name, initials, isLoggedIn} = useSession();
+  const { picture, name, initials, isLoggedIn } = useSession();
   const { signOut } = useNostrContext();
   const router = useRouter();
   const handleSignOut = useCallback(() => {
-    if(signOut) {
+    if (signOut) {
       signOut();
-      router.push('/');
+      router.push("/");
     }
   }, [router, signOut]);
 
@@ -107,47 +109,48 @@ export function Header() {
     <header className="flex h-14 w-full items-center justify-between bg-[#171B21] px-8">
       <MainNav items={HeaderConfig.mainNav} />
       <div className="hidden items-center md:inline">
-      <DropdownMenu>
-    <DropdownMenuTrigger asChild>
-      <div className="flex cursor-pointer">
-        <Avatar className="h-6 w-6">
-          <AvatarImage src={picture} />
-          <AvatarFallback>{initials}</AvatarFallback>
-        </Avatar>
-        <ChevronDown className="mt-1 h-4 w-4 hover:text-white/80" />
-      </div>
-    </DropdownMenuTrigger>
-    {isLoggedIn ? (<DropdownMenuContent className="w-56">
-      <DropdownMenuLabel>Signed in as {name}</DropdownMenuLabel>
-      <DropdownMenuSeparator />
-      <DropdownMenuGroup>
-        {PrimaryGitInfo?.map((item) => (
-          <DropdownMenuItem key={item.title}>
-            <span>{item.title}</span>
-          </DropdownMenuItem>
-        ))}
-        <DropdownMenuSeparator />
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <div className="flex cursor-pointer">
+              <Avatar className="h-6 w-6">
+                <AvatarImage src={picture} />
+                <AvatarFallback>{initials}</AvatarFallback>
+              </Avatar>
+              <ChevronDown className="mt-1 h-4 w-4 hover:text-white/80" />
+            </div>
+          </DropdownMenuTrigger>
+          {isLoggedIn ? (
+            <DropdownMenuContent className="w-56">
+              <DropdownMenuLabel>Signed in as {name}</DropdownMenuLabel>
+              <DropdownMenuSeparator />
+              <DropdownMenuGroup>
+                {PrimaryGitInfo?.map((item) => (
+                  <DropdownMenuItem key={item.title}>
+                    <span>{item.title}</span>
+                  </DropdownMenuItem>
+                ))}
+                <DropdownMenuSeparator />
 
-        {restGitInfo?.map((item) => (
-          <DropdownMenuItem key={item.title}>
-            <span>{item.title}</span>
-          </DropdownMenuItem>
-        ))}
-      </DropdownMenuGroup>
-      <DropdownMenuSeparator />
-      <DropdownMenuItem>
-        <Button onClick={handleSignOut}>Sign Out</Button>
-      </DropdownMenuItem>
-    </DropdownMenuContent>
-  ) : (
-    <DropdownMenuContent className="w-56">
-      <DropdownMenuItem>
-        <Link href="/login">Sign in</Link>
-      </DropdownMenuItem>
-    </DropdownMenuContent>
-  )}
-  </DropdownMenu>
-  </div>
-  </header>
+                {restGitInfo?.map((item) => (
+                  <DropdownMenuItem key={item.title}>
+                    <span>{item.title}</span>
+                  </DropdownMenuItem>
+                ))}
+              </DropdownMenuGroup>
+              <DropdownMenuSeparator />
+              <DropdownMenuItem>
+                <Button onClick={handleSignOut}>Sign Out</Button>
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          ) : (
+            <DropdownMenuContent className="w-56">
+              <DropdownMenuItem>
+                <Link href="/login">Sign in</Link>
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          )}
+        </DropdownMenu>
+      </div>
+    </header>
   );
 }
